Multiple changes in response to code review feedback.
Change the Attr mode string reading to be more clear.

Fix typos and clarify the function of the "force" parameter for the
remove functions to be just for ignoring when a file/directory doesn't
eixst.

Make the "remove" and "rmdirs" functions both throw Qfs::ENOENT when the
file/directory isn't found. The "force_remove" method then catches just
specifically this error.

Fix more typos